refactor: 프레젠테이션 Response DTO 금액 타입 Long → BigDecimal 변경
- PartialSettlementAvailableItemResponse: grossAmount, feeAmount, netAmount 변경
- PartialSettlementExecutionResponse: totalSalesAmount, feeAmount, finalSettlementAmount 변경

+++ settlement/src/main/java/com/example/settlement/presentation/dto/response/PartialSettlementExecutionResponse.java
@@ -3,6 +3,7 @@ package com.example.settlement.presentation.dto.response;
 import com.example.settlement.application.dto.PartialSettlementExecutionResult;
 import com.example.settlement.domain.enumtype.SettlementStatus;
 import com.example.settlement.domain.enumtype.SettlementType;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -14,9 +15,9 @@ public record PartialSettlementExecutionResponse(
         SettlementType settlementType,
         SettlementStatus settlementStatus,
         int settlementItemCount,
-        Long totalSalesAmount,
-        Long feeAmount,
-        Long finalSettlementAmount
+        BigDecimal totalSalesAmount,
+        BigDecimal feeAmount,
+        BigDecimal finalSettlementAmount
 ) {
 
     public static PartialSettlementExecutionResponse from(PartialSettlementExecutionResult result) {
